requirements/pypi.txt
# Django
django

# Model
django-ordered-model

# REST
djangorestframework
djangorestframework-filters
